pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calcifer"
version = "0.1.0"
description = "Local-first governance-first AI agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "openai",
  "python-dotenv",
  "pydantic",
  "requests",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["calcifer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
